+++ packages/frameworks/angular/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @qualcomm-ui/angular
 
+## 1.7.0 (2025/11/30)
+
+### Features
+
+- angular header-bar
+
+### Bug Fixes
+
+- context token providers in nav-item
+- use buttons QDS props rather than extending them
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/qds-core]
+
+## 1.6.0 (2025/11/30)
+
+### Features
+
+- wire highlightMatchingText in angular combobox
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/angular-core]
+
 ## 1.5.0 (2025/11/25)
 
 ### Features